+++ client/main.js
@@ -34,6 +34,7 @@ import '/imports/ui/pages/accounts/reset/reset.js';
 import '/imports/ui/pages/accounts/reset/resetSent.js';
 import '/imports/ui/pages/accounts/reset/resetPassword.js';
 import '/imports/ui/pages/accounts/reset/resetSuccess.js';
+import '/imports/ui/pages/accounts/pausedUser.js';
 
 
 // Planning
