+++ client/scss/_loaders.scss
@@ -201,6 +201,7 @@
 	.resetPassword,
 	.resetSuccess,
 	.billingError,
+	.pausedUser,
 	.loading-signing-out {
 		.frame-one {
 			.loading-container {
