+++ imports/ui/layouts/app.html
@@ -28,5 +28,5 @@
 		</div>
 	</div>
 
-	<span class="dis-lg-none">{{> navbar}}</span>
+	{{#if userId}}<span class="dis-lg-none">{{> navbar}}</span>{{/if}}
 </template>
