+++ imports/ui/pages/accounts/verify/verifySent.html
@@ -6,6 +6,6 @@
 		<p class="m-tn-t-3 m-tn-b-21 m-tn-rl-36">We sent you an email with a verification link.</p>
 	</main>
 	{{#unless Template.subscriptionsReady}}
-		{{> loading subbar=true}}
+		{{> loading}}
 	{{/unless}}
 </template>
