+++ imports/ui/pages/accounts/verify/verifySuccess.html
@@ -7,6 +7,6 @@
 		<a class="btn" href="{{pathFor 'studentsList'}}">Sign In</a>
 	</main>
 	{{#unless Template.subscriptionsReady}}
-		{{> loading subbar=true}}
+		{{> loading}}
 	{{/unless}}
 </template>
