+++ imports/ui/pages/reporting/reports/reportingSubjects.html
@@ -27,9 +27,9 @@
 						{{#if report.subjectsProgressVisible}}
 							<div class="meter meter-large m-tn-b-9">
 								<div class="meter-progress {{#if term.progressStatus}}meter-progress-primary{{/if}}" style="width: {{subject.progress}}%;">
-									<div class="meter-label">Term's Progress: {{subject.progress}}%</div>
+									<div class="meter-label">Subjects's Progress: {{subject.progress}}%</div>
 								</div>
-								<div class="meter-label">Term's Progress: {{subject.progress}}%</div>
+								<div class="meter-label">Subjects's Progress: {{subject.progress}}%</div>
 							</div>
 						{{/if}}
 						{{#if report.subjectsTimesVisible}}
