+++ imports/ui/pages/settings/billing/billingEdit.html
@@ -24,7 +24,7 @@
 			<input type="hidden" name="groupId" id="groupId" value="{{group._id}}">
 
 			<footer class="toolbar toolbar-bottom toolbar-gray">
-				<a href="#" class="js-cancel btn btn-gray-darker fl-tn-left">Cancel</a>
+				<a href="#" class="js-cancel btn btn-gray-darker fl-tn-left dis-sm-none">Cancel</a>
 				<button type="submit" class="js-submit fl-tn-right">Update</button>
 			</footer>
 		</form>
