+++ imports/ui/pages/settings/billing/billingInvoices.html
@@ -3,6 +3,7 @@
 	<main class="app-content app-content-preceeds-toolbar app-content-data">
 		<div class="list-info">
 			<ul class="list">
+				{{#if upcomingInvoices}}
 					<li class="list-item">
 						<a href="#" class="list-item-link dark-border p-tn-rl-9">
 							<div class="list-item-icon"><i class="txt-warning fss-circle"></i></div> 
@@ -14,6 +15,7 @@
 							</div>
 						</a>
 					</li>
+				{{/if}}
 
 				{{#each invoice in invoices.data}}
 					<li class="list-item">
